+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ChangeLog
 ---------
 
+### 1.11.1
+
+- fix(js): fix #2439, `filterBy` cannot filter array keys
+- fix(js): fix #2424, from html with checkbox bug
+- fix(js): fix #2385, checkbox render bug with formatter
+- fix(js): fix #750, showRow and hideRow bug
+- feat(js): fix #2414, add `paginationLoop` option
+- feat(cookie extension): fix #2386, add `getCookies` method
+
 ### 1.11.0
 
 - fix(js): fix cardVisible doesn't work bug
@@ -21,6 +30,7 @@ ChangeLog
 - fix(locale): add formatAllRows in template locale
 - fix(filter-control extension): add check for null values on existsOptionInSelectControl
 - fix(filter-control extension): fix show-clear button bug
+- fix(editable extesion): fix editable formatter error when refreshOptions
 - feat(js): add support for transfer from rowspan / colspan table
 - feat(js): add data variable to post-body event
 - feat(js): add `buttonsClass` option
@@ -38,6 +48,7 @@ ChangeLog
 - feat(angular extension): add constant in order to get it from angular scope
 - feat(export extension): add `formatExport` locale
 - feat(multiple-sort extension): add `formatSortOrders` option
+- feat(multiple-sort extension): support pagination server
 - refactor(filter-control extension): refactor the filterDataType method
 - refactor(filter-control extension): adding all unique values to select control and performance improvements
 - refactor(extension): refactor filter cookies extension to avoid dbcalls